Finish

+++ MusicSite/Models/Reviews/ReviewAdd.cs
@@ -16,7 +16,7 @@ namespace MusicSite.Models.Reviews
         [Display(Name ="Ocena")]
         [Required(ErrorMessage ="Nie można zostawić pustej oceny")]
         [Range(0, 5, ErrorMessage = "Ocena spoza zakresu")]
-        public byte rating { get; set; }
+        public int rating { get; set; }
         public string reviewText { get; set; }
     }
 }
